+++ includes/Entity/Character/ICharacter.hpp
@@ -8,7 +8,7 @@ class				ICharacter : public AEntity<T>
 {
 public:
   virtual eEntityType		getType() const = 0;
-  virtual IEntity<T>		*clone(int x, int y) = 0;
+  virtual IEntity<T>		*clone(double x, double y) = 0;
   virtual ~ICharacter() {};
 };
