+++ includes/Entity/Character/IPlayer.hpp
@@ -9,7 +9,7 @@ class				IPlayer : public ICharacter<T>
 public:
   virtual ~IPlayer() {};
   virtual eEntityType		getType() const = 0;
-  virtual IEntity<T>		*clone(int x, int y) = 0;
+  virtual IEntity<T>		*clone(double x, double y) = 0;
   virtual int			getBombStock() const = 0;
   virtual void			setBombStock(int nb) = 0;
 };
