+++ includes/Entity/Environment/IFloor.hpp
@@ -8,7 +8,7 @@ class				IFloor : public IEnvironment<T>
 {
 public:
   virtual eEntityType		getType() const = 0;
-  virtual IEntity<T>		*clone(int x, int y) = 0;
+  virtual IEntity<T>		*clone(double x, double y) = 0;
   virtual			~IFloor() {};
 };
